feat: split dashboard by user type
