+++ frontend/src/app/dashboard/artist/__tests__/RequestSortFilter.test.tsx
@@ -12,7 +12,7 @@ jest.mock('@/lib/api');
 jest.mock('@/contexts/AuthContext');
 jest.mock('next/navigation', () => ({
   useRouter: jest.fn(),
-  usePathname: jest.fn(() => '/dashboard'),
+  usePathname: jest.fn(() => '/dashboard/artist'),
 }));
 
 describe('DashboardPage booking request sort and filter', () => {
